Added test app

+++ Charger.ino
@@ -52,7 +52,7 @@ float Thermistor(float RawADC) {
 	float Temp;  // Dual-Purpose variable to save space.
 
 	Resistance = pad * ((1023.0 / RawADC) - 1);
-	Temp = log(Resistance); // Saving the Log(resistance) so not to calculate  it 4 times later
+	Temp = log((float)Resistance); // Saving the Log(resistance) so not to calculate  it 4 times later
 	Temp = 1 / (0.001129148 + (0.000234125 * Temp) + (0.0000000876741 * Temp * Temp * Temp));
 	Temp = Temp - 273.15;  // Convert Kelvin to Celsius
